+++ src/musiscape/__init__.py
@@ -68,4 +68,4 @@ def __dir__():
     return sorted(__all__)
 
 
-__version__ = "0.10.0"
+__version__ = "0.11.0"